requirements.txt
numpy==1.24.0
faiss-gpu==1.7.2
deepface>=0.0.75
tqdm>=4.64.1
pickle-mixin>=1.0.2
tensorflow>=2.10
